% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cad.R
\docType{data}
\name{cad}
\alias{cad}
\title{CAD Object of a Steel Panel}
\format{
A \code{mesh3d} object (list) with three main components:
\describe{
\item{vb}{A 4 x 267 matrix containing vertex coordinates in homogeneous form.}
\item{it}{A 3 x 343 matrix containing triangle indices.}
\item{normals}{A 4 x 267 matrix containing vertex normals.}
}
}
\source{
Theoretical CAD model of a steel panel.
}
\usage{
cad
}
\description{
This object corresponds to a portion of the theoretical CAD model of a steel
panel, stored as a \code{mesh3d} object.
}
\keyword{datasets}
