% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimControl-package.R
\docType{package}
\name{dimControl-package}
\alias{dimControl}
\alias{dimControl-package}
\title{dimControl: Tools for Geometric Processing and Dimensional Analysis}
\description{
A collection of tools for geometric processing, reconstruction, and dimensional
analysis of three-dimensional data.
}
\section{Main functionality}{


\code{dimControl} provides functionality for point cloud and mesh processing, boundary
extraction, angle estimation, component detection, surface reconstruction, and
geometric analysis.
}

\section{Getting started}{


Work in progress
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/modes-cemi/dimControl}
  \item \url{https://modes-cemi.github.io/dimControl/}
  \item Report bugs at \url{https://github.com/modes-cemi/dimControl/issues}
}

}
\author{
\strong{Maintainer}: Nataly Romaris \email{nataly.romaris@udc.es}

Authors:
\itemize{
  \item Olamar Benavente
  \item Ruben Fernandez-Casal
  \item Salvador Naya
  \item Javier Tarrio-Saavedra
}

}
\keyword{internal}
